Pull out more functions into dateTime
