+++ assets/tests/components/propertiesPanel/ghostPropertiesPanel.test.tsx
@@ -3,9 +3,9 @@ import renderer from "react-test-renderer"
 import GhostPropertiesPanel from "../../../src/components/propertiesPanel/ghostPropertiesPanel"
 import { BlockWaiver, Ghost } from "../../../src/realtime"
 import { Route } from "../../../src/schedule"
-import * as blockWaiverBanner from "../../../src/components/propertiesPanel/blockWaiverBanner"
+import * as dateTime from "../../../src/util/dateTime"
 
-jest.spyOn(blockWaiverBanner, "nowEpochSeconds").mockImplementation(() => 81720)
+jest.spyOn(dateTime, "nowEpochSeconds").mockImplementation(() => 81720)
 
 // Enable feature flags
 jest.mock("../../../src/laboratoryFeatures", () => ({
