+++ assets/tests/components/propertiesPanel/vehiclePropertiesPanel.test.tsx
@@ -5,14 +5,12 @@ import VehiclePropertiesPanel from "../../../src/components/propertiesPanel/vehi
 import { HeadwaySpacing } from "../../../src/models/vehicleStatus"
 import { BlockWaiver, Vehicle } from "../../../src/realtime"
 import { Route } from "../../../src/schedule"
-import * as blockWaiverBanner from "../../../src/components/propertiesPanel/blockWaiverBanner"
-
-jest.spyOn(blockWaiverBanner, "nowEpochSeconds").mockImplementation(() => 81720)
 import * as dateTime from "../../../src/util/dateTime"
 
 jest
   .spyOn(dateTime, "now")
   .mockImplementation(() => new Date("2018-08-15T17:41:21.000Z"))
+jest.spyOn(dateTime, "nowEpochSeconds").mockImplementation(() => 81720)
 
 jest.spyOn(Date, "now").mockImplementation(() => 234000)
